Implementation of bussines endpoints and new improvements

+++ src/api/src/main/java/com/department/transportation/trip/statistics/api/OutTopZonesDto.java
@@ -26,8 +26,8 @@ public class OutTopZonesDto {
     private String zone;
 
     @JsonProperty("pu_total")
-    private Integer pickupTotal;
+    private Long pickupTotal;
 
     @JsonProperty("do_total")
-    private Integer dropOffTotal;
+    private Long dropOffTotal;
 }
